Use the same launcher for both mac OS variants.
Having a binary launcher for the x86_64 mac OS variant and a script
launcher for the arm64 variant is annoying, because they need to be
configured differently: the x86_64 binary launcher must be configured in
the `Info.plist` file, whereas the arm64 script launcher must be
configured in the script itself. This means we would need to give
different instructions to users if they want to, say, increase the
memory allocated to Protégé, depending on which variant they use.

It’s better to have a single launcher that works identically on both
architecture. Since the binary launcher is x86_64-specific, we use the
script launcher instead.

Now the only difference between the arm64 and the x86_64 variant is the
bundled JRE.